Crate twizzler_kernel

Modules

arch

clock

condvar

idcounter

image

initrd

interrupt

log

machine

memory

mutex

Implementation of a mutex that sleeps threads when there is contention.

obj

once

operations

panic

processor

sched

spinlock

syscall

thread

utils

Macros

emerglog

emerglogln

log

logln

Statics

TEST

Traits

BootInfo

A collection of information made available to the kernel by the bootloader or arch-dep modules.

Functions

idle_main

init_threading

kernel_main

thread_main

thread_user_main⚠