Struct twizzler_kernel::machine::pc::serial::SerialPort

pub struct SerialPort {
    port: u16,
}

Fields
port: u16
Implementations

impl SerialPort

const INT_EN: u16

const IID: u16

const DATA: u16

const FIFO_CTRL: u16

const LINE_CTRL: u16

const MODEM_CTRL: u16

const LINE_STS: u16

const MODEM_STS: u16

const SCRATCH: u16

pub unsafe fn new(port: u16) -> Self

Construct a new serial port.

Safety

The supplied port must be a correct, functioning serial port on the system.

pub unsafe fn write_reg(&self, reg: u16, val: u8)

Write register.

Safety

Must be a valid register in the serial port register space.

pub unsafe fn read_reg(&self, reg: u16) -> u8

Read register.

Safety

Must be a valid register in the serial port register space.

pub fn init(&mut self)

fn line_sts(&mut self) -> LineStsFlags

pub fn send(&mut self, byte: u8)

pub fn receive(&mut self) -> u8

pub fn has_pending(&mut self) -> bool

pub fn read_modem_status(&mut self) -> u8

pub fn read_iid(&mut self) -> u8