Module twizzler_kernel::memory::frame

Manage physical frames.

On kernel initialization, the system will call into init in this module to pass information about physical memory regions. Once that call completes, the physical frame allocator is ready for use. This has to happen before any fully-bootstrapped memory manager is ready to use. Note, though, that this module may have to perform memory allocation during initialization, so it’ll have to make use of the bootstrap memory allocator.

Physical frames are physical pages of memory, whose size depends on the architecture compiled for. A given physical frame can either be zeroed (that is, the physical memory the frame refers to contains only zeros), or it can be indeterminate. This distinction is maintained because it’s common that we need to allocate zero pages AND pages that will be immediately overwritten. Upon allocation, the caller can request a zeroed frame or an indeterminate frame. The allocator will try to reserve known-zero frames for allocations that request them.

Structs

BootFrameAllocator

Frame

A physical frame.

PhysicalFrameFlags

Flags to control the state of a physical frame. Also used by the alloc functions to indicate what kind of physical frame is being requested.

Functions

alloc_frame

Allocate a physical frame.

free_frame

Free a physical frame.

init

Initialize the global physical frame allocator.

try_alloc_frame

Try to allocate a physical frame. The flags argument is the same as in alloc_frame. Returns None if no physical frame is available.