Function twizzler_kernel::memory::frame::alloc_frame

pub fn alloc_frame(flags: PhysicalFrameFlags) -> Frame

Allocate a physical frame.

The flags argument allows one to control if the resulting frame is zeroed or not. Note that passing PhysicalFrameFlags::ZEROED guarantees that the returned frame is zeroed, but the converse is not true.

The returned frame will have its ZEROED flag cleared. In the future, this will probably change to reflect the correct state of the frame.

Panic

Will panic if out of physical memory. For this reason, you probably want to use try_alloc_frame.

Examples

let uninitialized_frame = alloc_frame(PhysicalFrameFlags::empty());
let zeroed_frame = alloc_frame(PhysicalFrameFlags::ZEROED);