Trait twizzler_kernel::BootInfo

pub trait BootInfo {
    fn memory_regions(&self) -> &'static [MemoryRegion];

    fn kernel_image_info(&self) -> (VirtAddr, usize);

    fn get_system_table(&self, table: BootInfoSystemTable) -> VirtAddr;

    fn get_modules(&self) -> &'static [BootModule];
}

A collection of information made available to the kernel by the bootloader or arch-dep modules.

Required methods

fn memory_regions(&self) -> &'static [MemoryRegion]

Return a static array of memory regions for the system.

fn kernel_image_info(&self) -> (VirtAddr, usize)

Return the address and length of the whole kernel image.

fn get_system_table(&self, table: BootInfoSystemTable) -> VirtAddr

Get a system table, the kinds available depend on the platform and architecture.

fn get_modules(&self) -> &'static [BootModule]

Get a static array of the modules loaded by the bootloader